added permissions for all wipe commands

+++ JeezyCore/src/main/java/de/jeezycore/commands/punishments/wipe/WipeAll.java
@@ -16,15 +16,16 @@ public class WipeAll implements CommandExecutor {
 
             if (cmd.getName().equalsIgnoreCase("wipeall") && args.length == 0) {
 
-                WipeSQL execute = new WipeSQL();
-                execute.wipeAll(p);
-
+                if (p.hasPermission("jeezy.core.punishments.wipe.all")) {
+                    WipeSQL execute = new WipeSQL();
+                    execute.wipeAll(p);
+                } else {
+                    p.sendMessage("No permission.");
+                }
             } else {
                 p.sendMessage("Usage: /wipeall");
             }
-
         }
-
         return true;
     }
 }
